re-organising aomic datagrabbers into sub-package
