--- junifer/datagrabber/aomic.py
@@ -1,354 +1,0 @@
-"""Provide concrete implementations for AOMIC1000 data access."""
-
-from itertools import product
-from pathlib import Path
-from typing import Dict, List, Tuple, Union
-
-from junifer.datagrabber import PatternDataladDataGrabber
-
-from ..api.decorators import register_datagrabber
-from ..utils import raise_error
-
-
-@register_datagrabber
-class DataladAOMICID1000(PatternDataladDataGrabber):
-    """Concrete implementation for pattern-based data fetching of AOMICID1000.
-
-    Parameters
-    ----------
-    datadir : str or Path, optional
-        The directory where the datalad dataset will be cloned. If None,
-        the datalad dataset will be cloned into a temporary directory
-        (default None).
-    **kwargs
-        Keyword arguments passed to superclass.
-
-    """
-
-    def __init__(
-        self,
-        datadir: Union[str, Path, None] = None,
-        **kwargs,
-    ) -> None:
-        """Initialize the class."""
-        # The types of data
-        types = [
-            "BOLD",
-            "BOLD_confounds",
-            "T1w",
-            "probseg_CSF",
-            "probseg_GM",
-            "probseg_WM",
-            "DWI",
-        ]
-        # The patterns
-        patterns = {
-            "BOLD": (
-                "derivatives/fmriprep/sub-{subject}/func/"
-                "sub-{subject}_task-moviewatching_"
-                "space-MNI152NLin2009cAsym_desc-preproc_bold.nii.gz"
-            ),
-            "BOLD_confounds": (
-                "derivatives/fmriprep/sub-{subject}/func/"
-                "sub-{subject}_task-moviewatching_"
-                "desc-confounds_regressors.tsv"
-            ),
-            "T1w": (
-                "derivatives/fmriprep/sub-{subject}/anat/"
-                "sub-{subject}_space-MNI152NLin2009cAsym_"
-                "desc-preproc_T1w.nii.gz"
-            ),
-            "probseg_CSF": (
-                "derivatives/fmriprep/sub-{subject}/anat/"
-                "sub-{subject}_space-MNI152NLin2009cAsym_label-"
-                "CSF_probseg.nii.gz"
-            ),
-            "probseg_GM": (
-                "derivatives/fmriprep/sub-{subject}/anat/"
-                "sub-{subject}_space-MNI152NLin2009cAsym_label-"
-                "GM_probseg.nii.gz"
-            ),
-            "probseg_WM": (
-                "derivatives/fmriprep/sub-{subject}/anat/"
-                "sub-{subject}_space-MNI152NLin2009cAsym_label-"
-                "WM_probseg.nii.gz"
-            ),
-            "DWI": (
-                "derivatives/dwipreproc/sub-{subject}/dwi/"
-                "sub-{subject}_desc-preproc_dwi.nii.gz"
-            ),
-        }
-        # The replacements
-        replacements = ["subject"]
-        uri = "https://github.com/OpenNeuroDatasets/ds003097.git"
-        super().__init__(
-            types=types,
-            datadir=datadir,
-            uri=uri,
-            patterns=patterns,
-            replacements=replacements,
-        )
-
-
-@register_datagrabber
-class DataladAOMICPIOP1(PatternDataladDataGrabber):
-    """Concrete implementation for pattern-based data fetching of AOMICPIOP1.
-
-    Parameters
-    ----------
-    datadir : str or Path, optional
-        The directory where the datalad dataset will be cloned. If None,
-        the datalad dataset will be cloned into a temporary directory
-        (default None).
-    tasks : {"restingstate", "anticipation", "emomatching", "faces", "gstroop",
-        "workingmemory"} or list of the options, optional
-        AOMIC PIOP1 task sessions. If None, all available task sessions are
-        selected (default None).
-    **kwargs
-        Keyword arguments passed to superclass.
-
-    """
-
-    def __init__(
-        self,
-        datadir: Union[str, Path, None] = None,
-        tasks: Union[str, List[str], None] = None,
-        **kwargs,
-    ) -> None:
-        """Initialize the class."""
-        # The types of data
-        types = [
-            "BOLD",
-            "BOLD_confounds",
-            "T1w",
-            "probseg_CSF",
-            "probseg_GM",
-            "probseg_WM",
-            "DWI",
-        ]
-
-        if isinstance(tasks, str):
-            tasks = [tasks]
-
-        all_tasks = [
-            "restingstate",
-            "anticipation",
-            "emomatching",
-            "faces",
-            "gstroop",
-            "workingmemory",
-        ]
-
-        if tasks is None:
-            tasks = all_tasks
-        else:
-            for t in tasks:
-                if t not in all_tasks:
-                    raise_error(
-                        f"{t} is not a valid task in the AOMIC PIOP1"
-                        " dataset!"
-                    )
-
-        self.tasks = tasks
-
-        patterns = {
-            "BOLD": (
-                "derivatives/fmriprep/sub-{subject}/func/"
-                "sub-{subject}_task-{task}_"
-                "space-MNI152NLin2009cAsym_desc-preproc_bold.nii.gz"
-            ),
-            "BOLD_confounds": (
-                "derivatives/fmriprep/sub-{subject}/func/"
-                "sub-{subject}_task-{task}_"
-                "desc-confounds_regressors.tsv"
-            ),
-            "T1w": (
-                "derivatives/fmriprep/sub-{subject}/anat/"
-                "sub-{subject}_space-MNI152NLin2009cAsym_"
-                "desc-preproc_T1w.nii.gz"
-            ),
-            "probseg_CSF": (
-                "derivatives/fmriprep/sub-{subject}/anat/"
-                "sub-{subject}_space-MNI152NLin2009cAsym_label-"
-                "CSF_probseg.nii.gz"
-            ),
-            "probseg_GM": (
-                "derivatives/fmriprep/sub-{subject}/anat/"
-                "sub-{subject}_space-MNI152NLin2009cAsym_label-"
-                "GM_probseg.nii.gz"
-            ),
-            "probseg_WM": (
-                "derivatives/fmriprep/sub-{subject}/anat/"
-                "sub-{subject}_space-MNI152NLin2009cAsym_label-"
-                "WM_probseg.nii.gz"
-            ),
-            "DWI": (
-                "derivatives/dwipreproc/sub-{subject}/dwi/"
-                "sub-{subject}_desc-preproc_dwi.nii.gz"
-            ),
-        }
-        uri = "https://github.com/OpenNeuroDatasets/ds002785"
-        replacements = ["subject", "task"]
-        super().__init__(
-            types=types,
-            datadir=datadir,
-            uri=uri,
-            patterns=patterns,
-            replacements=replacements,
-        )
-
-    def __getitem__(self, element: Tuple[str, str]) -> Dict[str, Path]:
-        """Index one element in the dataset.
-
-        Parameters
-        ----------
-        element : tuple of str
-            The element to be indexed. First element in the tuple is the
-            subject, second element is the task.
-
-        Returns
-        -------
-        out : dict
-            Dictionary of paths for each type of data required for the
-            specified element.
-
-        """
-        sub, task = element
-
-        # depending on task 'acquisition is different'
-        task_acqs = {
-            "anticipation": "seq",
-            "emomatching": "seq",
-            "faces": "mb3",
-            "gstroop": "seq",
-            "restingstate": "mb3",
-            "workingmemory": "seq",
-        }
-        acq = task_acqs[task]
-        new_task = f"{task}_acq-{acq}"
-
-        out = super().__getitem__((sub, new_task))
-        out["meta"]["element"] = {"subject": sub, "task": task}
-        return out
-
-    def get_elements(self) -> List:
-        """Implement fetching list of subjects in the dataset.
-
-        Returns
-        -------
-        elements : list of str
-            The list of subjects in the dataset.
-
-        """
-        subjects = [f"{x:04d}" for x in range(1, 217)]
-        elems = []
-        for subject, task in product(subjects, self.tasks):
-            elems.append((subject, task))
-
-        return elems
-
-
-@register_datagrabber
-class DataladAOMICPIOP2(PatternDataladDataGrabber):
-    """Concrete implementation for pattern-based data fetching of AOMICPIOP2.
-
-    Parameters
-    ----------
-    datadir : str or Path, optional
-        The directory where the datalad dataset will be cloned. If None,
-        the datalad dataset will be cloned into a temporary directory
-        (default None).
-    tasks : {"restingstate", "stopsignal", "emomatching", "workingmemory"} or
-        list of the options, optional
-        AOMIC PIOP2 task sessions. If None, all available task sessions are
-        selected (default None).
-    **kwargs
-        Keyword arguments passed to superclass.
-
-    """
-
-    def __init__(
-        self,
-        datadir: Union[str, Path, None] = None,
-        tasks: Union[str, List[str], None] = None,
-        **kwargs,
-    ) -> None:
-        """Initialize the class."""
-        # The types of data
-        types = [
-            "BOLD",
-            "BOLD_confounds",
-            "T1w",
-            "probseg_CSF",
-            "probseg_GM",
-            "probseg_WM",
-            "DWI",
-        ]
-
-        if isinstance(tasks, str):
-            tasks = [tasks]
-
-        all_tasks = [
-            "restingstate",
-            "emomatching",
-            "workingmemory",
-            "stopsignal",
-        ]
-
-        if tasks is None:
-            tasks = all_tasks
-        else:
-            for t in tasks:
-                if t not in all_tasks:
-                    raise_error(
-                        f"{t} is not a valid task in the AOMIC PIOP2"
-                        " dataset!"
-                    )
-
-        self.tasks = tasks
-
-        patterns = {
-            "BOLD": (
-                "derivatives/fmriprep/sub-{subject}/func/"
-                "sub-{subject}_task-{task}_acq-seq_"
-                "space-MNI152NLin2009cAsym_desc-preproc_bold.nii.gz"
-            ),
-            "BOLD_confounds": (
-                "derivatives/fmriprep/sub-{subject}/func/"
-                "sub-{subject}_task-{task}_acq-seq_"
-                "desc-confounds_regressors.tsv"
-            ),
-            "T1w": (
-                "derivatives/fmriprep/sub-{subject}/anat/"
-                "sub-{subject}_space-MNI152NLin2009cAsym_"
-                "desc-preproc_T1w.nii.gz"
-            ),
-            "probseg_CSF": (
-                "derivatives/fmriprep/sub-{subject}/anat/"
-                "sub-{subject}_space-MNI152NLin2009cAsym_label-"
-                "CSF_probseg.nii.gz"
-            ),
-            "probseg_GM": (
-                "derivatives/fmriprep/sub-{subject}/anat/"
-                "sub-{subject}_space-MNI152NLin2009cAsym_label-"
-                "GM_probseg.nii.gz"
-            ),
-            "probseg_WM": (
-                "derivatives/fmriprep/sub-{subject}/anat/"
-                "sub-{subject}_space-MNI152NLin2009cAsym_label-"
-                "WM_probseg.nii.gz"
-            ),
-            "DWI": (
-                "derivatives/dwipreproc/sub-{subject}/dwi/"
-                "sub-{subject}_desc-preproc_dwi.nii.gz"
-            ),
-        }
-        uri = "https://github.com/OpenNeuroDatasets/ds002790"
-        replacements = ["subject", "task"]
-        super().__init__(
-            types=types,
-            datadir=datadir,
-            uri=uri,
-            patterns=patterns,
-            replacements=replacements,
-        )
