+++ junifer/datagrabber/aomic/__init__.py
@@ -1,0 +1,9 @@
+"""Provide imports for datagrabber 'aomic' sub-package."""
+
+# Authors: Federico Raimondo <f.raimondo@fz-juelich.de>
+#          Leonard Sasse <l.sasse@fz-juelich.de>
+# License: AGPL
+
+from .id1000 import DataladAOMICID1000
+from .piop1 import DataladAOMICPIOP1
+from .piop2 import DataladAOMICPIOP2
